requirements.txt
python-dotenv
requests
enocean
redis
RPi.GPIO